+++ app/src/main/java/ru/pyply/games/points/db/Player.java
@@ -3,10 +3,10 @@ package ru.pyply.games.points.db;
 import java.io.Serializable;
 
 public class Player implements Serializable {
-    public final long id;
+    public final int id;
     public final String name;
 
-    public Player(long id, String name) {
+    public Player(int id, String name) {
         this.id = id;
         this.name = name;
     }
